fix benchmarking test script
