+++ csrc/redis/resp.h
@@ -12,9 +12,76 @@
 #include <thread>
 #include <vector>
 
-enum class Op : uint8_t;
-struct Request;
-struct Completion;
+enum class Op : uint8_t { BATCH_TILE_GET, BATCH_TILE_SET, BATCH_TILE_EXISTS };
+
+/* shared communication state between threads executing a single batch operation
+all threads need to complete before the completion is sent */
+/* tiling refers to dividing work for batched operations between threads
+beforehand */
+struct BatchState {
+  // the number of tiles left to finish
+  // we can only send a completion when the last tile is finished
+  std::atomic<uint32_t> remaining_tiles{0};
+  std::atomic<bool> any_failed{false};
+
+  std::mutex err_mu;
+  std::string first_error;
+
+  // for batch exists, store the boolean results (0/1)
+  // IMPORTANT: not vector<bool> due to concurrent write data race
+  std::vector<uint8_t> exists_results;
+
+  // track batch operation type to avoid fragile req.op checks
+  Op batch_op;
+};
+
+/*
+LIFETIME GUARANTEE:
+We have a strict assumption that Python will NOT clean up any buffer memory
+before all C++ operations finish. This is guaranteed by the Python-side design
+where the caller holds references to all buffers until drain_completions()
+returns the corresponding future_id. Therefore, we do NOT need to track
+buf_owner references or acquire the GIL to prevent premature cleanup.
+We can safely use raw pointers extracted under the GIL without additional
+lifetime management on the C++ side.
+*/
+
+struct Request {
+  // the completion also has a future_id
+  // the caller is responsible for matching the request to the completion
+  uint64_t future_id = 0;
+  Op op;
+
+  // all operations use the batched structure (even single-item operations
+  // are treated as batches of size 1)
+  std::vector<std::string> keys;
+  std::vector<void*> buf_ptrs;
+  std::vector<size_t> buf_lens;
+
+  // shared batch state between threads executing a single batch operation
+  // so that they can coordinate when to send the completion
+  std::shared_ptr<BatchState> batch;
+
+  // for batch exists tiles, track which indices this tile is responsible for
+  size_t start_idx = 0;
+};
+
+struct Completion {
+  // the request also has a future_id
+  // the caller is responsible for matching the completion to the request
+  uint64_t future_id = 0;
+
+  // did the operation succeed?
+  bool ok = true;
+  // for EXISTS only (no result in the completion for SET and GET)
+  bool result_bool = false;
+
+  // for batch EXISTS, store multiple boolean results as bytes (0/1)
+  std::vector<uint8_t> result_bytes;
+
+  // error string if operation failed
+  std::string error;
+};
 
 class MultiRESPClient {
  public:
